refactor: tripmember 패키지 분리 (domain module)
