+++ api/src/main/java/com/tago/api/trip/application/TripService.java
@@ -5,9 +5,9 @@ import com.tago.api.trip.dto.response.TripStatusResponse;
 import com.tago.domain.member.domain.Member;
 import com.tago.domain.member.domain.vo.Gender;
 import com.tago.domain.trip.domain.Trip;
-import com.tago.domain.trip.domain.TripMember;
+import com.tago.domain.tripmember.domain.TripMember;
 import com.tago.domain.trip.dto.TripPreviewDto;
-import com.tago.domain.trip.repository.TripMemberRepository;
+import com.tago.domain.tripmember.repository.TripMemberRepository;
 import com.tago.domain.trip.service.TripQueryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -15,7 +15,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
