+++ api/src/main/java/com/tago/api/tripmember/application/TripMemberService.java
@@ -2,16 +2,15 @@ package com.tago.api.tripmember.application;
 
 
 import com.tago.api.tripmember.dto.response.TripMemberJoinResponse;
+import com.tago.domain.member.domain.Member;
 import com.tago.domain.member.exception.MaxMemberLimitException;
+import com.tago.domain.member.service.MemberQueryService;
 import com.tago.domain.trip.domain.Trip;
-import com.tago.domain.trip.domain.TripMember;
-import com.tago.domain.trip.service.TripCommandService;
-import com.tago.domain.trip.service.TripMemberCommandService;
-import com.tago.domain.trip.service.TripMemberCreateService;
+import com.tago.domain.tripmember.domain.TripMember;
+import com.tago.domain.tripmember.service.TripMemberCreateService;
 import com.tago.domain.trip.service.TripQueryService;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -20,15 +19,17 @@ public class TripMemberService {
 
     private final TripMemberCreateService tripMemberCreateService;
     private final TripQueryService tripQueryService;
+    private final MemberQueryService memberQueryService;
 
     @Transactional
     public TripMemberJoinResponse joinTrip(Long tripId, Long memberId){
         Trip trip = tripQueryService.findById(tripId);
+        Member member = memberQueryService.findById(memberId);
+
         canJoinTrip(trip);
         trip.incrementCurrentMember();
 
-        TripMember tripMember = tripMemberCreateService.create(tripId, memberId);
-
+        TripMember tripMember = tripMemberCreateService.create(trip, member);
 
         return new TripMemberJoinResponse(tripMember.getId());
     }
