+++ api/src/main/java/com/tago/api/tripmember/dto/response/TripMemberJoinResponse.java
@@ -1,5 +1,4 @@
 package com.tago.api.tripmember.dto.response;
-import com.tago.domain.trip.domain.TripMember;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
