+++ domain/src/main/java/com/tago/domain/tripmember/repository/TripMemberCustomRepository.java
@@ -1,0 +1,4 @@
+package com.tago.domain.tripmember.repository;
+
+public interface TripMemberCustomRepository {
+}
