+++ domain/src/main/java/com/tago/domain/tripmember/repository/TripMemberRepository.java
@@ -1,6 +1,6 @@
-package com.tago.domain.trip.repository;
+package com.tago.domain.tripmember.repository;
 
-import com.tago.domain.trip.domain.TripMember;
+import com.tago.domain.tripmember.domain.TripMember;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
